Moved logged in user function into its own service
