+++ src/main/java/uk/gov/hmcts/reform/prl/services/DraftAnOrderService.java
@@ -202,6 +202,7 @@ public class DraftAnOrderService {
     private final ElementUtils elementUtils;
     private final ObjectMapper objectMapper;
     private final ManageOrderService manageOrderService;
+    private final LoggedInUserService loggedInUserService;
     private final DgsService dgsService;
     private final UserService userService;
     private final DocumentLanguageService documentLanguageService;
@@ -225,7 +226,7 @@ public class DraftAnOrderService {
     private final WelshCourtEmail welshCourtEmail;
 
     public List<Element<DraftOrder>> generateDraftOrderCollection(CaseData caseData, String authorisation) {
-        String loggedInUserType = manageOrderService.getLoggedInUserType(authorisation);
+        String loggedInUserType = loggedInUserService.getLoggedInUserType(authorisation);
         List<Element<DraftOrder>> draftOrderList = new ArrayList<>();
         Element<DraftOrder> orderDetails = element(getCurrentOrderDetails(caseData, loggedInUserType, authorisation));
         //By default all the hearing will be option 1 (dateReservedWithListAssit) as per ticket PRL-4766
@@ -263,7 +264,7 @@ public class DraftAnOrderService {
     public Map<String, Object> getDraftOrderDynamicList(CaseData caseData,
                                                         String eventId,
                                                         String authorisation) {
-        String loggedInUserType = manageOrderService.getLoggedInUserType(authorisation);
+        String loggedInUserType = loggedInUserService.getLoggedInUserType(authorisation);
         Map<String, Object> caseDataMap = new HashMap<>();
         List<Element<DraftOrder>> supportedDraftOrderList = new ArrayList<>();
         caseData.getDraftOrderCollection().forEach(
@@ -355,7 +356,7 @@ public class DraftAnOrderService {
         List<Element<DraftOrder>> draftOrderCollection = caseData.getDraftOrderCollection();
         UUID selectedOrderId = elementUtils.getDynamicListSelectedValue(
             caseData.getDraftOrdersDynamicList(), objectMapper);
-        String loggedInUserType = manageOrderService.getLoggedInUserType(authorisation);
+        String loggedInUserType = loggedInUserService.getLoggedInUserType(authorisation);
         updatedCaseData.put(CASE_TYPE_OF_APPLICATION, CaseUtils.getCaseTypeOfApplication(caseData));
         for (Element<DraftOrder> e : caseData.getDraftOrderCollection()) {
             DraftOrder draftOrder = e.getValue();
@@ -464,7 +465,7 @@ public class DraftAnOrderService {
     }
 
     private Element<OrderDetails> convertDraftOrderToFinal(String auth, CaseData caseData, DraftOrder draftOrder, String eventId) {
-        String loggedInUserType = manageOrderService.getLoggedInUserType(auth);
+        String loggedInUserType = loggedInUserService.getLoggedInUserType(auth);
         OrderDetails orderDetails = getOrderDetails(
             caseData,
             draftOrder,
@@ -1071,7 +1072,7 @@ public class DraftAnOrderService {
 
     public Map<String, Object> updateDraftOrderCollection(CaseData caseData, String authorisation, String eventId, String draftOrderId) {
         List<Element<DraftOrder>> draftOrderCollection = caseData.getDraftOrderCollection();
-        String loggedInUserType = manageOrderService.getLoggedInUserType(authorisation);
+        String loggedInUserType = loggedInUserService.getLoggedInUserType(authorisation);
         UUID selectedOrderId;
         if (StringUtils.isEmpty(draftOrderId)) {
             if (Event.EDIT_RETURNED_ORDER.getId().equalsIgnoreCase(eventId)) {
