+++ src/main/java/uk/gov/hmcts/reform/prl/services/c100respondentsolicitor/C100RespondentSolicitorService.java
@@ -66,7 +66,7 @@ import uk.gov.hmcts.reform.prl.models.language.DocumentLanguage;
 import uk.gov.hmcts.reform.prl.services.ApplicationsTabService;
 import uk.gov.hmcts.reform.prl.services.ConfidentialityC8RefugeService;
 import uk.gov.hmcts.reform.prl.services.DocumentLanguageService;
-import uk.gov.hmcts.reform.prl.services.ManageOrderService;
+import uk.gov.hmcts.reform.prl.services.LoggedInUserService;
 import uk.gov.hmcts.reform.prl.services.OrganisationService;
 import uk.gov.hmcts.reform.prl.services.RespondentAllegationOfHarmService;
 import uk.gov.hmcts.reform.prl.services.SystemUserService;
@@ -163,7 +163,7 @@ public class C100RespondentSolicitorService {
     private final DocumentGenService documentGenService;
     private final ResponseSubmitChecker responseSubmitChecker;
     private final ApplicationsTabService applicationsTabService;
-    private final ManageOrderService manageOrderService;
+    private final LoggedInUserService loggedInUserService;
     private final SystemUserService systemUserService;
     private final ConfidentialDetailsMapper confidentialDetailsMapper;
     private final OrganisationService organisationService;
@@ -1252,7 +1252,7 @@ public class C100RespondentSolicitorService {
         if (dataMap.containsKey(IS_CONFIDENTIAL_DATA_PRESENT)) {
 
             //prl-6790 - getting user-role and adding to datamap
-            dataMap.put("loggedInUserRole", manageOrderService.getLoggedInUserType(authorisation));
+            dataMap.put("loggedInUserRole", loggedInUserService.getLoggedInUserType(authorisation));
 
             DocumentLanguage documentLanguage = documentLanguageService.docGenerateLang(caseData);
             if (documentLanguage.isGenWelsh()) {
