+++ src/main/java/uk/gov/hmcts/reform/prl/services/citizen/CitizenResponseService.java
@@ -30,7 +30,7 @@ import uk.gov.hmcts.reform.prl.models.documents.Document;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.language.DocumentLanguage;
 import uk.gov.hmcts.reform.prl.services.DocumentLanguageService;
-import uk.gov.hmcts.reform.prl.services.ManageOrderService;
+import uk.gov.hmcts.reform.prl.services.LoggedInUserService;
 import uk.gov.hmcts.reform.prl.services.c100respondentsolicitor.C100RespondentSolicitorService;
 import uk.gov.hmcts.reform.prl.services.document.DocumentGenService;
 import uk.gov.hmcts.reform.prl.services.tab.alltabs.AllTabServiceImpl;
@@ -80,7 +80,7 @@ public class CitizenResponseService {
     private final DocumentGenService documentGenService;
     private final DocumentLanguageService documentLanguageService;
     private final CitizenPartyDetailsMapper citizenPartyDetailsMapper;
-    private final ManageOrderService manageOrderService;
+    private final LoggedInUserService loggedInUserService;
 
     DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("ddmmyyyy");
     public static final String C1A_RESPONSE = "C1Aresponse";
@@ -398,7 +398,7 @@ public class CitizenResponseService {
         if (dataMap.containsKey(IS_CONFIDENTIAL_DATA_PRESENT)) {
             int partyIndex = caseData.getRespondents().indexOf(partyDetailsElement);
             //prl-6790 - getting user-role and adding to datamap
-            dataMap.put("loggedInUserRole", manageOrderService.getLoggedInUserType(authorisation));
+            dataMap.put("loggedInUserRole", loggedInUserService.getLoggedInUserType(authorisation));
 
             Document c8FinalDocument = documentGenService.generateSingleDocument(
                     authorisation,
