+++ src/test/java/uk/gov/hmcts/reform/prl/services/DgsServiceTest.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.reform.prl.services;
 
 import feign.FeignException;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.function.ThrowingRunnable;
 import org.junit.runner.RunWith;
@@ -35,7 +34,6 @@ import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-@Ignore
 public class DgsServiceTest {
 
     @InjectMocks
@@ -52,6 +50,8 @@ public class DgsServiceTest {
 
     @Mock
     private HearingDataService hearingDataService;
+    @Mock
+    private LoggedInUserService loggedInUserService;
 
     public static final String authToken = "Bearer TestAuthToken";
     public static final String PRL_DRAFT_TEMPLATE = "FL-DIV-GOR-ENG-00062.docx";
