+++ src/test/java/uk/gov/hmcts/reform/prl/services/citizen/CitizenResponseServiceTest.java
@@ -34,6 +34,7 @@ import uk.gov.hmcts.reform.prl.models.documents.Document;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.language.DocumentLanguage;
 import uk.gov.hmcts.reform.prl.services.DocumentLanguageService;
+import uk.gov.hmcts.reform.prl.services.LoggedInUserService;
 import uk.gov.hmcts.reform.prl.services.ManageOrderService;
 import uk.gov.hmcts.reform.prl.services.c100respondentsolicitor.C100RespondentSolicitorService;
 import uk.gov.hmcts.reform.prl.services.document.DocumentGenService;
@@ -63,6 +64,9 @@ public class CitizenResponseServiceTest {
     CitizenResponseService citizenResponseService;
     @Mock
     ObjectMapper objectMapper;
+
+    @Mock
+    LoggedInUserService loggedInUserService;
     @Mock
     CcdCoreCaseDataService ccdCoreCaseDataService;
     @Mock
